De-copupled the creation of indexes from table creation, making it possible to perform initial table loads without the overhead of updating additional indexes.
Also added a new bulk load mechanism via the export of a pipe separated file which is then copied to the SQL server and bulk inserted locally. In testing this mechanism is up to 20 times faster than the previous mechanism. The new bulk upload requires xfServer to be installed and licensed on the database server, and is not currently supported from OpenVMS environments.